# Improve perf by converting sparse grads to dense.