Benchmark results: I ran on a Z840 with a Titan V. I added a `sys.exit(0)` after 1200 steps to end early.  I ran 20 iterations, both with and without this commit, at batch sizes 2048 and 65536.

```
python ncf_main.py --model_dir=$HOME/ncf_model_dir_gpu_1 --data_dir=$HOME/ncf_data_dir_gpu_1 --dataset ml-20m --hooks= --num_gpus -1 --clean --train_epochs 14 --batch_size $BATCH_SIZE --eval_batch_size 65536 --layers 256,256,128,64 --num_factors 64 --hr_threshold 0.99 --ml_perf --use_synthetic_data --hooks=examplespersecondhook
```

| Batch Size | Examples/sec without commit | Examples/sec with commit | Percent improvement |
| --- | --- | --- | --- |
| 2048 | 288774.5 | 354672.6 | 22.81% |
| 65536 | 3962970.1 | 5105452.2 | 28.83 % |

I tested this does not affect convergence. Running with hyperparameters batch_size=65536, learning_rate=0.00281936, beta1=0.0490852, beta2=0.965095, epsilon=1.23757e-08, I reached a HR of 63.5% 92/100 times without this commit, and 75/80 times with this commit.